Documents/practice/kaggle/12041310/Digit Classifier/requirements.txt
streamlit
Pillow
torch
numpy
